backend_legacy/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=openblas
bottleneck=1.3.5=py310h96f19d2_0
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.05.30=hca03da5_0
catalogue=2.0.7=py310hca03da5_0
certifi=2023.5.7=py310hca03da5_0
cffi=1.15.1=py310h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310hca03da5_0
colorama=0.4.6=py310hca03da5_0
confection=0.0.4=py310h33ce5c2_0
cryptography=39.0.1=py310h834c97f_0
cymem=2.0.6=py310hc377ac9_0
cython-blis=0.7.9=py310hbda83bc_0
dataclasses=0.8=pyh6d0b6a4_7
dnspython=2.3.0=pyhd8ed1ab_0
en-core-web-sm=3.5.0=pypi_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.9.0=py310hca03da5_0
flask=2.2.2=py310hca03da5_0
flask_cors=3.0.10=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
fsspec=2023.4.0=py310hca03da5_0
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
gnutls=3.6.15=h887c41c_0
huggingface_hub=0.14.1=py310hca03da5_0
icu=68.1=hc377ac9_0
idna=3.4=py310hca03da5_0
importlib-metadata=6.0.0=py310hca03da5_0
importlib_metadata=6.0.0=hd3eb1b0_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py310hca03da5_0
joblib=1.2.0=py310hca03da5_0
jpeg=9e=h80987f9_1
lame=3.100=h1a28f6b_0
langcodes=3.3.0=pyhd3eb1b0_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.4=h80987f9_0
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libtasn1=4.19.0=h80987f9_0
libtiff=4.5.0=h313beb8_2
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
libxml2=2.10.3=h372ba2a_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py310h1a28f6b_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.2.1=pypi_0
murmurhash=1.0.7=py310hc377ac9_0
ncurses=6.4=h313beb8_0
nettle=3.7.3=h84b5d62_1
networkx=2.8.4=py310hca03da5_1
nltk=3.7=pyhd3eb1b0_0
numexpr=2.8.4=py310hecc3335_1
numpy=1.22.3=py310hdb36b11_0
numpy-base=1.22.3=py310h5e3e9f0_0
openh264=1.8.0=h98b2900_0
openssl=1.1.1u=h1a28f6b_0
packaging=23.0=py310hca03da5_0
pandas=1.5.3=py310h46d7db6_0
pathy=0.10.1=py310hca03da5_0
pillow=9.0.1=py310h4d1bdd5_0
pip=23.0.1=py310hca03da5_0
pooch=1.4.0=pyhd3eb1b0_0
preshed=3.0.6=py310hc377ac9_0
protobuf=3.20.3=py310h313beb8_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.2=py310h1a28f6b_0
pyjwt=2.4.0=py310hca03da5_0
pymongo=3.12.0=py310hc377ac9_0
pyopenssl=23.0.0=py310hca03da5_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.0=hbdb9e5c_5
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=2.0.1=py3.10_0
pytz=2022.7=py310hca03da5_0
pyyaml=6.0=py310h80987f9_1
readline=8.2=h1a28f6b_0
regex=2022.7.9=py310h1a28f6b_0
requests=2.29.0=py310hca03da5_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-learn=1.0.2=py310h59830a0_1
scipy=1.10.1=py310h20cbe94_1
setuptools=67.8.0=py310hca03da5_0
shellingham=1.5.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py310hca03da5_0
spacy=3.5.3=py310hb6aeb05_0
spacy-legacy=3.0.12=py310hca03da5_0
spacy-loggers=1.0.4=py310hca03da5_0
sqlite=3.41.2=h80987f9_0
srsly=2.4.6=py310h313beb8_0
sympy=1.11.1=py310hca03da5_0
thinc=8.1.10=py310hb6aeb05_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.11.4=py310h5c5695e_1
torchaudio=2.0.2=py310_cpu
torchvision=0.15.2=py310_cpu
tqdm=4.65.0=py310h33ce5c2_0
transformers=4.28.1=py_0
typer=0.4.1=py310hca03da5_0
typing-extensions=4.5.0=py310hca03da5_0
typing_extensions=4.5.0=py310hca03da5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py310hca03da5_0
vadersentiment=3.3.2=pyhd8ed1ab_0
wasabi=0.9.1=py310hca03da5_0
werkzeug=2.2.3=py310hca03da5_0
wheel=0.38.4=py310hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xz=5.4.2=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zipp=3.11.0=py310hca03da5_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
